prep for graph analysis

+++ lib/common/include/common_time.hpp
@@ -8,10 +8,10 @@
 
 namespace ssp4cpp::common::time
 {
-    int constexpr milliseconds_per_seconds  = 1000;
-    int constexpr microseconds_per_seconds  = 1000000;
-    int constexpr nanoseconds_per_seconds  =  1000000000;
+    uint64_t constexpr milliseconds_per_seconds  = 1000;
+    uint64_t constexpr microseconds_per_seconds  = 1000000;
+    uint64_t constexpr nanoseconds_per_seconds  =  1000000000;
     
-    int constexpr seconds_per_hour =  3600;
-    int constexpr hours_per_day = 24;
+    uint64_t constexpr seconds_per_hour =  3600;
+    uint64_t constexpr hours_per_day = 24;
 }
